Split tests/runner.py. NFC
tests/runner.py is the a python entry point which means it ends being
registered as `__main__` and not `runner`.  This means that when the
other modules import `runner` python imports the file a second time.

This means that we have two differing copies of runner.py in memory at
once with potentially different state.  It easy to see this today if you
add `print("IMPORTING RUNNER")` to the top of the file.

This change split runner.py into the actual test runner code and the
common code which is shared between the different tests suites.
Splitting it up this way seems better, regardless of the issue immediate
double-import issue.
